backend/requirements.txt
Django==3.2.23
djangorestframework==3.14.0
bcrypt==4.1.2
python-dotenv==1.0.1
django-cors-headers==4.3.0
sqlparse==0.4.4 